server/requirements.txt
api==0.0.7
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
nose==1.3.7
points==0.4.0
postgres==3.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
